pyproject.toml
[project]
name = "music-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepaudio-x>=0.4.0",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "pyyaml>=6.0.3",
    "sounddevice>=0.5.5",
]
